# openapi/fc_openapi.yaml
openapi: 3.0.1
info:
  title: Eclipse XFSC Federated Catalogue
  description: REST API of the XFSC catalogue
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
  - url: https://fc-server.xfsc.org
    description: some future test environment
components:
  responses:
...
    ServerError:
      description: >-
        May contain hints how to solve the error or indicate what went wrong at
        the server. Must not outline any information about the internal
        structure of the server.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ClientError:
      description: >-
        May contain hints how to solve the error or indicate what was wrong in
        the request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required:
        - code
        - message
    Statement:
      type: object
      properties:
        statement:
          type: string
          example: Match (m:Movie) where m.released > 2000 RETURN m
        parameters:
          $ref: '#/components/schemas/Parameters'
      required:
        - statement
#      discriminator:
#        propertyName: type
#        mapping:
#          as: "#/components/schemas/AnnotatedStatement"  
    Parameters: 
      type: object
      additionalProperties:
        type: object
      example:
        name: "John"
        limit: 10
        offset: 5
    Result:
      type: object
      additionalProperties:
        type: object
    QueryLanguage:
      type: string
      enum:
        - OPENCYPHER
        - SPARQL
        - GRAPHQL
      default: OPENCYPHER
    AnnotatedStatement:
      allOf:
        - $ref: '#/components/schemas/Statement'
        - type: object
          properties:
            servers:
              type: array
              uniqueItems: true
              items:
                type: string
                description: known catalog servers already queried in distributed search scenario
            annotations:
              type: object
              additionalProperties:
                type: object
                description: additional properties to tune query execution
              example:
                queryLanguage: SPARQL
                timeout: 30
                withTotalCount: false
...
    ResultPage:
      type: object
      properties:
        totalCount:
          type: integer
...
    Results:
      allOf:
        - $ref: '#/components/schemas/ResultPage'
        - type: object
          properties:
            items: 
              type: array
              items: 
                $ref: '#/components/schemas/Result'
...
  securitySchemes:
    jwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /query: #!! Suggestion: rename to query, and allow only a single query. If multiple queries are ever added in the future, add an endpoint queries. -> Julius: ok
    get:
      tags:
        - Query
      summary: Retrieve an HTML website to send openCypher queries to the Catalogue
      operationId: querywebsite
      security:
        - jwt: []
      responses:
        '200':
          description: HTML document that contains a query field for openCypher queries.
          content:
            text/html:
              schema:
                type: string
        '400':
          $ref: '#/components/responses/ClientError'
        '500':
          $ref: '#/components/responses/ServerError'
    post:
      tags:
        - Query
      summary: Send a query to the Catalogue
      operationId: query
      parameters:
        - in: query
          name: queryLanguage
          schema:
            $ref: '#/components/schemas/QueryLanguage'
          required: false
        - in: query
          name: timeout
          schema:
            type: integer
            default: 5
          required: false
        - in: query
          name: withTotalCount
          schema:
            type: boolean
            default: true
          required: false
      security:
        - jwt: []
      requestBody:
        description: >-
          JSON object to send queries. Use "application/json" for openCypher
          queries. A Catalogue may also support the other content types
          depending on its supported query languages but only "application/json"
          is mandatory.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Statement'
      responses:
        '200':
          description: successful query
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Results'
        '400': ## Malformed Message: The receveived request cannot be processed, either because its syntax is ## incorrect or forbidden query clauses are used. For instance, it is not allowed to manipulate data ## through the query endpoint
          $ref: '#/components/responses/ClientError'
        '408': ## Query Timeout: the query took longer than the configured timeout
          ## interval. The client needs to rewrite the query so it can be
          ## processed faster.
          $ref: '#/components/responses/ServerError'
        '500':
          $ref: '#/components/responses/ServerError'
...